@@ -1,30 +1,28 @@
-/*
- * Created by Kultala Aki on 4/24/21 9:34 AM
- * Copyright (c) 2021. All rights reserved.
- * Last modified 4/2/21 1:22 PM
- */
-
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'com.google.gms.google-services'
-apply plugin: 'kotlin-android-extensions'
-apply plugin: 'com.google.firebase.crashlytics'
+plugins {
+    alias(libs.plugins.android.application)
+    alias(libs.plugins.google.services)
+    alias(libs.plugins.google.firebase.crashlytics)
+    alias(libs.plugins.google.firebase.firebase.perf)
+}
 
 android {
-    compileSdkVersion 30
+    namespace 'kultalaaki.vpkapuri'
+    compileSdk 34
 
     defaultConfig {
         applicationId "kultalaaki.vpkapuri"
-        minSdkVersion 21
-        targetSdkVersion 30
-        multiDexEnabled true
-        versionCode 110 // Update just before pushing update 109 is live
-        versionName "3.9" // Update just before pushing update 3.81 is live
+        minSdk 26
+        targetSdk 34
+        versionCode 111
+        versionName "3.9.51"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+
     buildTypes {
         release {
-            minifyEnabled true
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
     compileOptions {
