@@ -8,7 +8,6 @@ package kultalaaki.vpkapuri;
 
 
 import android.Manifest;
-import android.annotation.TargetApi;
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.app.NotificationManager;
@@ -32,15 +31,18 @@ import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
 import android.preference.RingtonePreference;
 import android.provider.Settings;
-import androidx.annotation.NonNull;
-import androidx.annotation.RequiresApi;
-import androidx.core.app.ActivityCompat;
 import android.text.TextUtils;
 import android.view.MenuItem;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+
 import java.util.ArrayList;
 import java.util.List;
+
+import kultalaaki.vpkapuri.misc.SeekBarPreference;
+import kultalaaki.vpkapuri.util.Constants;
 
 /**
  * A {@link PreferenceActivity} that presents a set of application settings. On
@@ -154,7 +156,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
     //Oman menun testailu
 
 
-
     /**
      * {@inheritDoc}
      */
@@ -167,7 +168,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
      * {@inheritDoc}
      */
     @Override
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public void onBuildHeaders(List<Header> target) {
         loadHeadersFromResource(R.xml.pref_headers, target);
     }
@@ -188,7 +188,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
      * This fragment shows general pref_sounds only. It is used when the
      * activity is showing a two-pane settings UI.
      */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public static class GeneralPreferenceFragment extends PreferenceFragment {
 
 
@@ -224,12 +223,23 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             bindPreferenceSummaryToValue(findPreference("halyvastaanotto8"));
             bindPreferenceSummaryToValue(findPreference("halyvastaanotto9"));
             bindPreferenceSummaryToValue(findPreference("halyvastaanotto10"));
-            bindPreferenceSummaryToValue(findPreference("avainsana1"));
-            bindPreferenceSummaryToValue(findPreference("avainsana2"));
-            bindPreferenceSummaryToValue(findPreference("avainsana3"));
-            bindPreferenceSummaryToValue(findPreference("avainsana4"));
-            bindPreferenceSummaryToValue(findPreference("avainsana5"));
+            bindPreferenceSummaryToValue(findPreference("unit1"));
+            bindPreferenceSummaryToValue(findPreference("unit2"));
+            bindPreferenceSummaryToValue(findPreference("unit3"));
+            bindPreferenceSummaryToValue(findPreference("unit4"));
+            bindPreferenceSummaryToValue(findPreference("unit5"));
+            bindPreferenceSummaryToValue(findPreference("unit6"));
+            bindPreferenceSummaryToValue(findPreference("unit7"));
+            bindPreferenceSummaryToValue(findPreference("unit8"));
+            bindPreferenceSummaryToValue(findPreference("unit9"));
+            bindPreferenceSummaryToValue(findPreference("unit10"));
             bindPreferenceSummaryToValue(findPreference("AlarmCounterTime"));
+
+            bindPreferenceSummaryToValue(findPreference("vapepanumber1"));
+            bindPreferenceSummaryToValue(findPreference("vapepanumber2"));
+            bindPreferenceSummaryToValue(findPreference("vapepanumber3"));
+            bindPreferenceSummaryToValue(findPreference("vapepanumber4"));
+            bindPreferenceSummaryToValue(findPreference("vapepanumber5"));
 
             /*
             Preference button = findPreference(getString(R.string.Aseta_numerot));
@@ -256,7 +266,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
         }
 
 
-
         @Override
         public boolean onOptionsItemSelected(MenuItem item) {
             int id = item.getItemId();
@@ -272,7 +281,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
      * This fragment shows notification pref_sounds only. It is used when the
      * activity is showing a two-pane settings UI.
      */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public static class NotificationPreferenceFragment extends PreferenceFragment implements SharedPreferences.OnSharedPreferenceChangeListener {
 
         private SeekBarPreference _seekBarPref;
@@ -288,15 +296,14 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
 
             //Do Not Disturb allowed if menu doesn't open
             Preference pref = getPreferenceManager().findPreference("DoNotDisturb");
-            if(pref != null) {
+            if (pref != null) {
                 pref.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
                         NotificationManager notificationManager =
                                 (NotificationManager) getActivity().getApplicationContext().getSystemService(Context.NOTIFICATION_SERVICE);
-                        if(notificationManager != null) {
-                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M
-                                    && !notificationManager.isNotificationPolicyAccessGranted()) {
+                        if (notificationManager != null) {
+                            if (!notificationManager.isNotificationPolicyAccessGranted()) {
                                 Intent intent = new Intent(android.provider.Settings.ACTION_NOTIFICATION_POLICY_ACCESS_SETTINGS);
                                 startActivity(intent);
                                 return true;
@@ -321,25 +328,23 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             // updated to reflect the new value, per the Android Design
             // guidelines.
             //bindPreferenceSummaryToValue(findPreference("notifications_new_message_ringtone"));
-            bindPreferenceSummaryToValue(findPreference("ringtone"));
+            bindPreferenceSummaryToValue(findPreference("ringtone_rescue"));
             bindPreferenceSummaryToValue(findPreference("stopTime"));
             bindPreferenceSummaryToValue(findPreference("vibrate_pattern"));
 
 
         }
 
-        public long[] genVibratorPattern( float intensity, long duration )
-        {
-            float dutyCycle = Math.abs( ( intensity * 2.0f ) - 1.0f );
-            long hWidth = (long) ( dutyCycle * ( duration - 1 ) ) + 1;
+        public long[] genVibratorPattern(float intensity, long duration) {
+            float dutyCycle = Math.abs((intensity * 2.0f) - 1.0f);
+            long hWidth = (long) (dutyCycle * (duration - 1)) + 1;
             long lWidth = dutyCycle == 1.0f ? 0 : 1;
 
-            int pulseCount = (int) ( 2.0f * ( (float) duration / (float) ( hWidth + lWidth ) ) );
-            long[] pattern = new long[ pulseCount ];
+            int pulseCount = (int) (2.0f * ((float) duration / (float) (hWidth + lWidth)));
+            long[] pattern = new long[pulseCount];
 
-            for( int i = 0; i < pulseCount; i++ )
-            {
-                pattern[i] = intensity < 0.5f ? ( i % 2 == 0 ? hWidth : lWidth ) : ( i % 2 == 0 ? lWidth : hWidth );
+            for (int i = 0; i < pulseCount; i++) {
+                pattern[i] = intensity < 0.5f ? (i % 2 == 0 ? hWidth : lWidth) : (i % 2 == 0 ? lWidth : hWidth);
             }
 
             return pattern;
@@ -351,45 +356,43 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             // Set seekbar summary :
             //SharedPreferences pref = PreferenceManager.getDefaultSharedPreferences(this.getActivity());
             //int rad = sharedPreferences.getInt("SEEKBAR_VALUE", 50);
-            if(key.equals("vibrate_pattern")) {
+            if (key.equals("vibrate_pattern")) {
 
                 ListPreference vibratePattern = (ListPreference) findPreference("vibrate_pattern");
-                if(vibratePattern != null) {
+                if (vibratePattern != null) {
                     //CharSequence vibrateText = vibratePattern.getEntry();
                     //String vibrateValue = vibratePattern.getValue();
                     int vibrateValue = Integer.parseInt(vibratePattern.getValue());
                     long[] pattern = new long[]{};
                     int[] amplitude = new int[]{};
 
-                    if(vibrateValue == 0) {
+                    if (vibrateValue == 0) {
                         pattern = Constants.PULSE_PATTERN;
                         amplitude = Constants.PULSE_AMPLITUDE;
                     } else if (vibrateValue == 1) {
                         pattern = Constants.HURRY_PATTERN;
                         amplitude = Constants.HURRY_AMPLITUDE;
-                    } else if(vibrateValue == 2) {
+                    } else if (vibrateValue == 2) {
                         pattern = Constants.SLOW_PATTERN;
                         amplitude = Constants.SLOW_AMPLITUDE;
-                    } else if(vibrateValue == 3) {
+                    } else if (vibrateValue == 3) {
                         pattern = Constants.SOS_PATTERN;
                         amplitude = Constants.SOS_AMPLITUDE;
-                    } else if(vibrateValue == 4) {
+                    } else if (vibrateValue == 4) {
                         pattern = Constants.VIRVE_PATTERN;
                         amplitude = Constants.VIRVE_AMPLITUDE;
                     }
-                    if(Build.VERSION.SDK_INT >= 21) {
-                        if(viber != null && viber.hasVibrator()) {
-                            if(Build.VERSION.SDK_INT >= 26 && viber.hasAmplitudeControl()) {
-                                viber.vibrate(VibrationEffect.createWaveform(pattern, amplitude, -1));
-                            } else {
-                                viber.vibrate(pattern, -1);
-                            }
+                    if (viber != null && viber.hasVibrator()) {
+                        if (viber.hasAmplitudeControl()) {
+                            viber.vibrate(VibrationEffect.createWaveform(pattern, amplitude, -1));
+                        } else {
+                            viber.vibrate(VibrationEffect.createWaveform(pattern, -1));
                         }
                     }
                 }
             }
             Activity activity = getActivity();
-            if(activity != null) {
+            if (activity != null) {
                 int radius = PreferenceManager.getDefaultSharedPreferences(this.getActivity()).getInt("SEEKBAR_VALUE", 50);
                 _seekBarPref.setSummary(this.getString(R.string.settings_summary).replace("$1", "" + radius));
             }
@@ -411,13 +414,16 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
      * This fragment shows data and sync pref_sounds only. It is used when the
      * activity is showing a two-pane settings UI.
      */
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public static class DataSyncPreferenceFragment extends PreferenceFragment {
 
-        private String[] permissionsSms = {Manifest.permission.READ_CALL_LOG, Manifest.permission.READ_EXTERNAL_STORAGE,
-                Manifest.permission.READ_PHONE_STATE, Manifest.permission.RECEIVE_SMS, Manifest.permission.SEND_SMS};
-
-        //, Manifest.permission.CALL_PHONE
+        private final String[] permissionsSms = {
+                Manifest.permission.POST_NOTIFICATIONS,
+                Manifest.permission.SYSTEM_ALERT_WINDOW,
+                Manifest.permission.READ_EXTERNAL_STORAGE,
+                Manifest.permission.RECEIVE_SMS,
+                Manifest.permission.SEND_SMS,
+                Manifest.permission.READ_MEDIA_AUDIO,
+                Manifest.permission.REQUEST_COMPANION_START_FOREGROUND_SERVICES_FROM_BACKGROUND};
 
         @Override
         public void onCreate(Bundle savedInstanceState) {
@@ -426,24 +432,21 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
 
             //Battery optimization
             Preference prefi = getPreferenceManager().findPreference("batteryOptimization");
-            if(prefi != null) {
+            if (prefi != null) {
                 prefi.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                            Intent intent = new Intent(Settings.ACTION_IGNORE_BATTERY_OPTIMIZATION_SETTINGS);
-                            startActivity(intent);
-                            return true;
-                        }
+                        Intent intent = new Intent(Settings.ACTION_IGNORE_BATTERY_OPTIMIZATION_SETTINGS);
+                        startActivity(intent);
+                        return true;
 
-                        return false;
                     }
                 });
             }
 
             //Opening app from background
             Preference prefOpenWindow = getPreferenceManager().findPreference("windowOperation");
-            if(prefOpenWindow != null) {
+            if (prefOpenWindow != null) {
                 prefOpenWindow.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
@@ -462,7 +465,7 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             }
 
             Preference prefTietosuoja = getPreferenceManager().findPreference("tietosuoja");
-            if(prefTietosuoja != null) {
+            if (prefTietosuoja != null) {
                 prefTietosuoja.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
@@ -476,8 +479,8 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
                 });
             }
 
-            Preference prefKayttoehdot= getPreferenceManager().findPreference("kayttoehdot");
-            if(prefKayttoehdot != null) {
+            Preference prefKayttoehdot = getPreferenceManager().findPreference("kayttoehdot");
+            if (prefKayttoehdot != null) {
                 prefKayttoehdot.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
@@ -491,8 +494,8 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
                 });
             }
 
-            Preference prefFacebook= getPreferenceManager().findPreference("facebook");
-            if(prefFacebook != null) {
+            Preference prefFacebook = getPreferenceManager().findPreference("facebook");
+            if (prefFacebook != null) {
                 prefFacebook.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
@@ -522,38 +525,33 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
 
             //Sovelluksen luvat
             Preference prefa = getPreferenceManager().findPreference("luvatApuri");
-            if(prefa != null) {
+            if (prefa != null) {
                 prefa.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
-                            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                                Intent intent = new Intent();
-                                intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-                                Uri uri = Uri.fromParts("package", getActivity().getPackageName(), null);
-                                intent.setData(uri);
-                                startActivity(intent);
-                                return true;
-                            }
+                        Intent intent = new Intent();
+                        intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
+                        Uri uri = Uri.fromParts("package", getActivity().getPackageName(), null);
+                        intent.setData(uri);
+                        startActivity(intent);
+                        return true;
 
-                        return false;
                     }
                 });
             }
 
             //Sovelluksen luvat SMS hälytyksiä varten
             Preference prefsms = getPreferenceManager().findPreference("luvatSMS");
-            if(prefsms != null) {
+            if (prefsms != null) {
                 prefsms.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
                     @Override
                     public boolean onPreferenceClick(Preference preference) {
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
-                            if(!arePermissionsEnabled()){
-                                // permissions granted, continue flow normally
-                                requestMultiplePermissions();
-                            }//else{
-                              //  requestMultiplePermissions();
-                            //}
-                        }
+                        if (!arePermissionsEnabled()) {
+                            // permissions granted, continue flow normally
+                            requestMultiplePermissions();
+                        }//else{
+                        //  requestMultiplePermissions();
+                        //}
                         return true;
                     }
                 });
@@ -568,17 +566,15 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             //bindPreferenceSummaryToValue(findPreference("sync_frequency"));
         }
 
-        @RequiresApi(api = Build.VERSION_CODES.M)
-        private boolean arePermissionsEnabled(){
-            for(String permission : permissionsSms){
-                if(ActivityCompat.checkSelfPermission(getActivity(), permission) != PackageManager.PERMISSION_GRANTED)
+        private boolean arePermissionsEnabled() {
+            for (String permission : permissionsSms) {
+                if (ActivityCompat.checkSelfPermission(getActivity(), permission) != PackageManager.PERMISSION_GRANTED)
                     return false;
             }
             return true;
         }
 
-        @RequiresApi(api = Build.VERSION_CODES.M)
-        private void requestMultiplePermissions(){
+        private void requestMultiplePermissions() {
             List<String> remainingPermissions = new ArrayList<>();
             for (String permission : permissionsSms) {
                 if (ActivityCompat.checkSelfPermission(getActivity(), permission) != PackageManager.PERMISSION_GRANTED) {
@@ -590,15 +586,14 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             //requestPermissions(remainingPermissions.toArray(new String[remainingPermissions.size()]), 101);
         }
 
-        @TargetApi(Build.VERSION_CODES.M)
         @Override
         public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions,
                                                @NonNull int[] grantResults) {
             super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-            if(requestCode == 101){
-                for(int i=0;i<grantResults.length;i++){
-                    if(grantResults[i] != PackageManager.PERMISSION_GRANTED){
-                        if(shouldShowRequestPermissionRationale(permissions[i])){
+            if (requestCode == 101) {
+                for (int i = 0; i < grantResults.length; i++) {
+                    if (grantResults[i] != PackageManager.PERMISSION_GRANTED) {
+                        if (shouldShowRequestPermissionRationale(permissions[i])) {
                             new AlertDialog.Builder(getActivity())
                                     .setMessage("Annoitko tarvittavat luvat?")
                                     .setPositiveButton("Näytä uudestaan", new DialogInterface.OnClickListener() {
@@ -634,7 +629,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
         }
     }
 
-    @TargetApi(Build.VERSION_CODES.HONEYCOMB)
     public static class AsematauluPreferenceFragment extends PreferenceFragment {
 
 
@@ -1099,7 +1093,6 @@ public class SettingsActivity extends AppCompatPreferenceActivity {
             bindPreferenceSummaryToValue(findPreference("optional4_50"));
             bindPreferenceSummaryToValue(findPreference("optional5_50"));
         }
-
 
 
         @Override
